requirements.txt
fastapi
python-dotenv
lxml
tomli
uvicorn
requests